requirements.txt
requests
kafka-python
python-dotenv
pubg-python
influxdb-client